feat: allow plugins to set custom patterns for dangerous commands (#281)
